+++ flamenco/eve_settings.py
@@ -28,7 +28,6 @@ managers_schema = {
     # on how a job type looks like (in terms of tasks).
     'job_types': {
         'type': 'dict',
-        # TODO: will be renamed to 'keyschema' in Cerberus 1.0
         'keyschema': {  # name of the job type
             'type': 'string',
         },
@@ -37,7 +36,6 @@ managers_schema = {
             'schema': {
                 'vars': {
                     'type': 'dict',
-                    # TODO: will be renamed to 'keyschema' in Cerberus 1.0
                     'keyschema': {  # name of the variable
                         'type': 'string',
                     },
@@ -119,8 +117,14 @@ managers_schema = {
                 'type': 'integer',
             }
         }
-    }
-
+    },
+    # All the task types this Manager's Workers support.
+    'worker_task_types': {
+        'type': 'list',
+        'schema': {
+            'type': 'string',
+        },
+    },
 }
 
 jobs_schema = {
@@ -289,6 +293,8 @@ tasks_schema = {
             'blender-render',  # Rendering with Blender.
             'exr-merge',  # EXR merging, probably also happens with Blender.
             'file-management',  # removing directory trees, moving files around, etc.
+            'video-encoding',  # Running things through ffmpeg to produce videos.
+            'debug',  # Debugging stuff.
         ]
     },
     'commands': {